backend/requirements.txt
fastapi
httpx
python-dotenv
pydantic
uvicorn[Standard]
openai